requirements.txt
streamlit
pandas
pdfplumber
pypdf
openai
pydantic>=2.0
tenacity
python-dotenv
